# Speed up loading time entry list