--- a/src/ui/windows/TogglDesktop/TogglDesktop/ui/ViewModels/ViewModelExtensions.cs
+++ b/src/ui/windows/TogglDesktop/TogglDesktop/ui/ViewModels/ViewModelExtensions.cs
@@ -55,5 +55,30 @@ namespace TogglDesktop.ViewModels
                 item.Billable,
                 false);
         }
+
+        public static bool IsEqualTo(
+            this TimeEntryLabelViewModel timeEntryLabelViewModel,
+            Toggl.TogglTimeEntryView item)
+        {
+            if (timeEntryLabelViewModel == null) return false;
+            return timeEntryLabelViewModel.Description == item.Description
+                   && timeEntryLabelViewModel.Tags == item.Tags
+                   && timeEntryLabelViewModel.IsBillable == item.Billable
+                   && timeEntryLabelViewModel.ProjectLabel.IsEqualTo(item);
+        }
+
+        public static bool IsEqualTo(
+            this ProjectLabelViewModel projectLabelViewModel,
+            Toggl.TogglTimeEntryView item)
+        {
+            if (projectLabelViewModel == null) return false;
+            return projectLabelViewModel.ProjectName == item.ProjectLabel
+                   && projectLabelViewModel.TaskName == item.TaskLabel
+                   && projectLabelViewModel.ClientName == item.ClientLabel
+                   && projectLabelViewModel.ColorString == item.Color
+                   && projectLabelViewModel.WorkspaceName == item.WorkspaceName
+                   && projectLabelViewModel.ProjectInfo.ProjectId == item.PID
+                   && projectLabelViewModel.ProjectInfo.TaskId == item.TID;
+        }
     }
 }