### 📒 Description
Speed up loading time entry list by avoiding recreating and reassigning a time entry label view model if the time entry hasn't changed.

### 🕶️ Types of changes
 - **Bug fix** (non-breaking change which fixes an issue)

### 🤯 List of changes
- [x] compare an already loaded time entry label view model with the incoming time entry model
- [x] do not recreate a time entry label view model if it's already the same as the time entry model

### 👫 Relationships
Closes #4550.

### 🔎 Review hints
- Log in to an account with or create 200+ TEs
- Try opening/closing Edit view and expanding/collapsing TEs on `master` and on the current branch.
- Check if the performance improvements are tangible
- In Performance.cs line 12 set `debug` to `true`
- Try opening/closing Edit view and expanding/collapsing TEs on `master` and on the current branch.
- Look at the performance measurements in the logs


### Performance Measurements
Method: `OnTimeEntryList`
Input Data: 259 TEs, about 10% of which are grouped into 2-3 TEs.
Actions: Opening/Closing Edit view, Expanding/Collapsing grouped TEs
Before: 324ms - 545ms
After: 59ms - 110ms